+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 0xMiden/miden-node/pull/1194)).
 - Refresh `miden-base` dependencies.
 - Added `SyncTransactions` gRPC endpoint for retrieving transactions for specific accounts within a block range ([#1224](https://github.com/0xMiden/miden-node/pull/1224)).
+- Added `miden-network-monitor` binary for monitoring the Miden network ([#1217](https://github.com/0xMiden/miden-node/pull/1217)).
 
 ## v0.11.0 (2025-08-28)